review changes

+++ src/main/java/stsjorbsmod/powers/DamnationPower.java
@@ -27,15 +27,17 @@ public class DamnationPower extends CustomJorbsModPower implements NonStackableP
 
     @Override
     public void onUseCard(AbstractCard card, UseCardAction action) {
-        CardMetaUtils.destroyCardPermanently(card);
-        AbstractDungeon.transformCard(card, isAutoUpgrade, AbstractDungeon.miscRng);
-        AbstractDungeon.topLevelEffectsQueue.add(new ShowCardAndObtainEffect(AbstractDungeon.getTransformedCard(), Settings.WIDTH / 2.0F, Settings.HEIGHT / 2.0F, false));
-        addToBot(new RemoveSpecificPowerAction(owner, owner, this.ID));
+        if (!card.purgeOnUse) {
+            CardMetaUtils.destroyCardPermanently(card);
+            AbstractDungeon.transformCard(card, isAutoUpgrade, AbstractDungeon.miscRng);
+            AbstractDungeon.topLevelEffectsQueue.add(new ShowCardAndObtainEffect(AbstractDungeon.getTransformedCard(), Settings.WIDTH / 2.0F, Settings.HEIGHT / 2.0F, false));
+        }
+        addToTop(new RemoveSpecificPowerAction(owner, owner, this.ID));
     }
 
     @Override
     public void updateDescription() {
-        description = DESCRIPTIONS[0] + (isAutoUpgrade ? DESCRIPTIONS[2] : DESCRIPTIONS[1]);
+        description = isAutoUpgrade ? DESCRIPTIONS[0] : DESCRIPTIONS[1];
     }
 
     @Override
